pyproject.toml
[project]
name = "src"
version = "1.0.0"
description = ""
authors = [
    { name = "Dario", email = "dario@mail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "django-environ (>=0.12.0,<0.13.0)",
    "bugsnag (>=4.8.0,<5.0.0)",
    "django (>=5.2.8,<6.0.0)",
    "django-allauth[socialaccount] (>=65.13.1,<66.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "slack-sdk (>=3.39.0,<4.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "faker (>=39.0.0,<40.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "mysqlclient (>=2.2.8,<3.0.0)",
    "manticoresearch (>=11.0.0,<12.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "django-compressor (>=4.6.0,<5.0.0)",
    "geoip2 (>=5.2.0,<6.0.0)",
    "python-telegram-bot (>=22.7,<23.0)",
    "openai (>=2.30.0,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

